Add files via upload
The Galaxy_survey_lib_with_HST notebook compares the RA and Dec of different surveys to match up the galaxies and create tables with the information. The cosmos_spec_matching notebook takes in those tables to plot the spectra for the same galaxy from different surveys. I have uploaded the vUDS spectra and the necessary tables in the SC-SN-DATA dropbox folder. The rest of the data should be in the DATA dropbox folder.